delete unnecessary go file
